Add roads code to utils and move all roads constants over to a function in config.py that gets loaded in utils
