+++ gch4i/config.py
@@ -63,6 +63,45 @@ def load_global_file_names():
           Grid_area01_inputfile, Grid_area001_inputfile, Grid_state001_ansi_inputfile, \
           Grid_county001_ansi_inputfile)
 
+def load_road_globals():
+    raw_path = Path(V3_DATA_PATH) / "global/raw"
+    raw_roads_path = Path(V3_DATA_PATH) / "global/raw_roads"
+
+    road_file = str(raw_path / "tl_")
+    raw_road_file = str(raw_roads_path / "tl_")
+
+    task_outputs_path = Path(V3_DATA_PATH) / "global/raw_roads/task_outputs"
+
+    global_path = Path(V3_DATA_PATH) / "global"
+    gdf_state_files = str(global_path / "tl_2020_us_state/tl_2020_us_state.shp")
+
+    global_input_path = Path(V3_DATA_PATH.parent) / "GEPA_Source_code/Global_InputData"
+    state_ansi_path = str(global_input_path / "ANSI/ANSI_state.txt")
+
+    GEPA_Comb_Mob_path = Path(V3_DATA_PATH.parent) / "GEPA_Source_Code/GEPA_Combustion_Mobile/InputData"
+    State_vmt_file = str(GEPA_Comb_Mob_path / "vm2/vm2_")
+    State_vdf_file = str(GEPA_Comb_Mob_path / "vm4/vm4_")
+
+    # ANSI State Codes
+    #State_ANSI, name_dict = data_load_fn.load_state_ansi(state_ansi_path)[0:2]
+    State_ANSI, name_dict = load_state_ansi(state_ansi_path)[0:2]
+
+    State_ANSI['State_Num'] = State_ANSI.index + 1
+    state_mapping = State_ANSI.set_index('State_Num')['abbr']
+
+    #  Define local variables
+    start_year = min_year  # 2012 First year in emission timeseries
+    end_year = max_year    # 2022 Last year in emission timeseries
+    year_range = [*range(min_year, max_year+1, 1)]  # List of emission years
+    year_range_str = [str(i) for i in year_range]
+    num_years = len(year_range)
+
+    return (
+        raw_path, raw_roads_path, road_file, raw_road_file, task_outputs_path, global_path, 
+        gdf_state_files, global_input_path, state_ansi_path, GEPA_Comb_Mob_path, State_vmt_file, 
+        State_vdf_file, State_ANSI, name_dict, state_mapping, start_year, 
+        end_year, year_range, year_range_str, num_years 
+    )
 
 #Load grid cell areas (m2), and lat and lon values for 0.01x0.01 grid
 def load_area_map_001(Grid_area001_inputfile):
@@ -124,4 +163,6 @@ def load_state_ansi(state_ansi_inputfile):
     name_dict = State_ANSI.set_index('name')['ansi'].to_dict()
     
     return(State_ANSI, name_dict, abbr_dict)
+
+
